requirements.txt
# Core dependencies for the VQA BEiT‑3 project
torch>=2.0.0
transformers>=4.35.0
pandas>=2.0.0
numpy>=1.23.0
scikit-learn>=1.3.0
pyyaml>=6.0
pillow>=9.5.0
tqdm>=4.66.0